fourth commit

+++ service/src/main/java/ru/mail/denis/service/DTOmodels/BasketDTO.java
@@ -1,8 +1,9 @@
 package ru.mail.denis.service.DTOmodels;
 
-import ru.mail.denis.models.User;
+import ru.mail.denis.repositories.model.User;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * Created by user on 27.07.2017.
@@ -14,7 +15,7 @@ public class BasketDTO implements Serializable {
     private Integer bookId;
     private String bookName;
     private Integer bookQuantity;
-    private Double bookPrice;
+    private BigDecimal bookPrice;
     private User user;
 
     public Integer getBookId() {
@@ -49,11 +50,11 @@ public class BasketDTO implements Serializable {
         this.bookQuantity = bookQuantity;
     }
 
-    public Double getBookPrice() {
+    public BigDecimal getBookPrice() {
         return bookPrice;
     }
 
-    public void setBookPrice(Double bookPrice) {
+    public void setBookPrice(BigDecimal bookPrice) {
         this.bookPrice = bookPrice;
     }
